Make base_flat_file specs use a shared example group
